// angular-frontend/src/app/models/pagination.model.ts
export interface PageResponse<T> {
  content: T[];
  totalElements: number;
  totalPages: number;
  size: number;
  number: number;
  first: boolean;
  last: boolean;
  empty: boolean;
}

export interface PageRequest {
  page: number;
  size: number;
  sort?: string;
  title?: string;
  authorId?: number;
  genreId?: number;
  publicationYear?: number;
}